add composite geojson in test_data
